--- src/services/index.ts
@@ -1,11 +1,0 @@
-import { getBitcoinBalance } from './get-bitcoin-balance';
-import { getBitcoinDataPrices } from './get-bitcoin-data-prices';
-import { getBitcoinHistoricalPrice } from './get-bitcoin-historical-price';
-import { getBitcoinTransactions } from './get-bitcoin-transactions';
-
-export {
-	getBitcoinBalance,
-	getBitcoinDataPrices,
-	getBitcoinHistoricalPrice,
-	getBitcoinTransactions,
-};
