--- src/stores/index.ts
@@ -1,6 +1,0 @@
-import { useAppSettings } from './use-app-settings';
-import { useBitcoinDataPrices } from './use-bitcoin-data-prices';
-import { useBitcoinHistoricalPrice } from './use-bitcoin-historical-price';
-import { useUserData } from './use-user-data';
-
-export { useAppSettings, useBitcoinDataPrices, useBitcoinHistoricalPrice, useUserData };
