+++ src/types/GetBitcoinDataPricesType.ts
